Preview thumbnails will now be un-rotated before op
Explicitly specified the Dark theme for the app
Changed images to be centred within the left pane of each item viewer,
to look neater, and support in-place transforms
Removed RotateImage function from ItemViewer
Add GB localisation
Reworked main page XAML to be easier to follow and scale better with
very small screen sizes
Removed rotation preference from the manifest - the app will work in all
orientations
Added minimum supported width of 500 pixels
Replaced calls to ContainerFromIndex with ContainerFromItem, since I'm
moving the item reference around anyway

+++ JPG Spinner/MainPage.xaml.h
@@ -19,11 +19,11 @@ namespace JPG_Spinner
 
 		property BOOL TrimChecked;
 		property BOOL ProgressiveChecked;
-		property Platform::IBox<int>^ cts
+		property Platform::IBox<int64>^ cts
 		{
-			Platform::IBox<int>^ get()
+			Platform::IBox<int64>^ get()
 			{
-				return ref new Platform::Box<int>((int)&_cts);
+				return ref new Platform::Box<int64>((int64)&_cts);
 			}
 		};
